+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/RazorLanguageServer.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using MediatR;
 using Microsoft.AspNetCore.Razor.LanguageServer.AutoInsert;
 using Microsoft.AspNetCore.Razor.LanguageServer.Common;
+using Microsoft.AspNetCore.Razor.LanguageServer.Common.Serialization;
 using Microsoft.AspNetCore.Razor.LanguageServer.Completion;
 using Microsoft.AspNetCore.Razor.LanguageServer.Formatting;
 using Microsoft.AspNetCore.Razor.LanguageServer.Hover;
@@ -29,15 +30,6 @@ using ILanguageServer = OmniSharp.Extensions.LanguageServer.Server.ILanguageServ
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer
 {
-    public static class SerializerExtensions
-    {
-        public static void RegisterConverter(this Serializer serializer, Newtonsoft.Json.JsonConverter converter)
-        {
-            serializer.JsonSerializer.Converters.Add(converter);
-            serializer.Settings.Converters.Add(converter);
-        }
-    }
-
     public sealed class RazorLanguageServer
     {
         private RazorLanguageServer()
@@ -47,6 +39,7 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer
         public static Task<ILanguageServer> CreateAsync(Stream input, Stream output, Trace trace)
         {
             Serializer.Instance.RegisterConverter(SemanticTokensOrSemanticTokensEditsConverter.Instance);
+            Serializer.Instance.JsonSerializer.Converters.RegisterRazorConverters();
 
             ILanguageServer server = null;
             server = OmniSharp.Extensions.LanguageServer.Server.LanguageServer.PreInit(options =>
