+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/Models/SemanticTokenLegendParams.cs
@@ -5,7 +5,7 @@ using MediatR;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic.Models
 {
-    internal class SemanticTokenLegendParams : IRequest<SemanticTokenLegendResponse>, IBaseRequest
+    internal class SemanticTokenLegendParams : IRequest<SemanticTokensLegend>, IBaseRequest
     {
     }
 }
