+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Semantic/RazorSemanticTokenLegendEndpoint.cs
@@ -3,7 +3,6 @@
 
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Razor.LanguageServer.Semantic.Capabilities;
 using Microsoft.AspNetCore.Razor.LanguageServer.Semantic.Interfaces;
 using Microsoft.AspNetCore.Razor.LanguageServer.Semantic.Models;
 
@@ -11,16 +10,9 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Semantic
 {
     internal class RazorSemanticTokenLegendEndpoint : ISemanticTokenLegendHandler
     {
-        private SemanticTokenLegendCapability _capability;
-
-        public Task<SemanticTokenLegendResponse> Handle(SemanticTokenLegendParams request, CancellationToken cancellationToken)
+        public Task<SemanticTokensLegend> Handle(SemanticTokenLegendParams request, CancellationToken cancellationToken)
         {
             return Task.FromResult(SemanticTokenLegend.Instance);
-        }
-
-        public void SetCapability(SemanticTokenLegendCapability capability)
-        {
-            _capability = capability;
         }
     }
 }
