+++ src/Razor/src/Microsoft.AspNetCore.Razor.VSCode/src/Semantic/RazorDocumentSemanticTokensProvider.ts
@@ -15,7 +15,7 @@ export class RazorDocumentSemanticTokensProvider
         previousResultId: string,
         token: vscode.CancellationToken,
     ): Promise<vscode.SemanticTokens | vscode.SemanticTokensEdits | undefined> {
-        let semanticTokenResponse = await this.serviceClient.mapSemanticTokensEdit(LanguageKind.Razor, document.uri, previousResultId);
+        let semanticTokenResponse = await this.serviceClient.semanticTokensEdit(LanguageKind.Razor, document.uri, previousResultId);
 
         if (semanticTokenResponse instanceof vscode.SemanticTokens) {
             // However we're serializing into Uint32Array doesn't set byteLength, which is checked by some stuff under the covers.
@@ -32,7 +32,7 @@ export class RazorDocumentSemanticTokensProvider
         range: vscode.Range,
         token: vscode.CancellationToken,
     ): Promise<vscode.SemanticTokens | undefined> {
-        let semanticRangeResponse = await this.serviceClient.mapSemanticTokensRange(LanguageKind.Razor, document.uri, range);
+        let semanticRangeResponse = await this.serviceClient.semanticTokensRange(LanguageKind.Razor, document.uri, range);
 
         if (semanticRangeResponse) {
             // However we're serializing into Uint32Array doesn't set byteLength, which is checked by some stuff under the covers.
@@ -45,7 +45,7 @@ export class RazorDocumentSemanticTokensProvider
     }
 
     public async provideDocumentSemanticTokens(document: vscode.TextDocument, token: vscode.CancellationToken): Promise<vscode.SemanticTokens | undefined> {
-        let semanticTokenResponse = await this.serviceClient.mapSemanticTokens(LanguageKind.Razor, document.uri);
+        let semanticTokenResponse = await this.serviceClient.semanticTokens(LanguageKind.Razor, document.uri);
 
         if (semanticTokenResponse) {
             // However we're serializing into Uint32Array doesn't set byteLength, which is checked by some stuff under the covers.
