+++ src/Razor/test/testapps/LargeProject/LargeProject.csproj
@@ -1,9 +1,5 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
-  <PropertyGroup>
-    <RazorSdkDirectoryRoot>$(RazorSdkArtifactsDirectory)$(Configuration)\sdk-output\</RazorSdkDirectoryRoot>
-  </PropertyGroup>
-
   <!--
   The AspNetCore repo depends on AspNetCore-Tooling, and we want to avoid creating a cylical dependency between the two
   by adding a reference to Microsoft.AspNetCore.App. We'll instead simulate the presence of the reference.
